+++ bot/src/types.ts
@@ -1,3 +1,5 @@
+import { Context } from "grammy/web";
+
 export interface Env {
   BOT_TOKEN: string;
   API_URL: string;
@@ -5,4 +7,13 @@ export interface Env {
 
 export type HonoContext = {
   Bindings: Env;
+};
+
+export interface BotConfig {
+    botToken: string;
+    botApiUrl: string;
+}
+
+export type MyContext = Context & {
+    botConfig: BotConfig;
 };
